refactoring of layer_state classes
