+++ lib/painting/color_pick_painter.dart
@@ -15,11 +15,12 @@
  */
 
 import 'package:flutter/material.dart';
+import 'package:kpix/layer_states/drawing_layer_state.dart';
+import 'package:kpix/layer_states/layer_state.dart';
 import 'package:kpix/models/app_state.dart';
 import 'package:kpix/util/helper.dart';
 import 'package:kpix/painting/itool_painter.dart';
 import 'package:kpix/painting/kpix_painter.dart';
-import 'package:kpix/widgets/main/layer_widget.dart';
 
 class ColorPickPainter extends IToolPainter
 {
