+++ lib/painting/itool_painter.dart
@@ -17,6 +17,8 @@
 import 'dart:collection';
 import 'dart:math';
 import 'package:get_it/get_it.dart';
+import 'package:kpix/layer_states/drawing_layer_state.dart';
+import 'package:kpix/layer_states/layer_state.dart';
 import 'package:kpix/tool_options/line_options.dart';
 import 'package:kpix/util/helper.dart';
 import 'package:kpix/models/app_state.dart';
@@ -26,7 +28,6 @@ import 'package:kpix/managers/preference_manager.dart';
 import 'package:kpix/painting/shader_options.dart';
 import 'package:kpix/tool_options/pencil_options.dart';
 import 'package:kpix/util/typedefs.dart';
-import 'package:kpix/widgets/main/layer_widget.dart';
 
 class BorderCoordinateSetI
 {
