+++ lib/painting/kpix_painter.dart
@@ -20,6 +20,9 @@ import 'dart:ui' as ui;
 
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
+import 'package:kpix/layer_states/drawing_layer_state.dart';
+import 'package:kpix/layer_states/layer_state.dart';
+import 'package:kpix/layer_states/reference_layer_state.dart';
 import 'package:kpix/preferences/gui_preferences.dart';
 import 'package:kpix/util/helper.dart';
 import 'package:kpix/models/app_state.dart';
@@ -37,7 +40,6 @@ import 'package:kpix/painting/spray_can_painter.dart';
 import 'package:kpix/painting/stamp_painter.dart';
 import 'package:kpix/managers/preference_manager.dart';
 import 'package:kpix/util/typedefs.dart';
-import 'package:kpix/widgets/main/layer_widget.dart';
 
 class KPixPainterOptions
 {
