+++ lib/painting/line_painter.dart
@@ -16,6 +16,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
+import 'package:kpix/layer_states/layer_state.dart';
 import 'package:kpix/managers/hotkey_manager.dart';
 import 'package:kpix/util/helper.dart';
 import 'package:kpix/painting/itool_painter.dart';
@@ -24,7 +25,6 @@ import 'package:kpix/managers/preference_manager.dart';
 import 'package:kpix/tool_options/line_options.dart';
 import 'package:kpix/tool_options/pencil_options.dart';
 import 'package:kpix/util/typedefs.dart';
-import 'package:kpix/widgets/main/layer_widget.dart';
 
 class LinePainter extends IToolPainter
 {
