+++ lib/painting/pencil_painter.dart
@@ -17,6 +17,8 @@
 import 'dart:collection';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
+import 'package:kpix/layer_states/drawing_layer_state.dart';
+import 'package:kpix/layer_states/layer_state.dart';
 import 'package:kpix/managers/hotkey_manager.dart';
 import 'package:kpix/tool_options/line_options.dart';
 import 'package:kpix/util/helper.dart';
@@ -25,7 +27,6 @@ import 'package:kpix/painting/kpix_painter.dart';
 import 'package:kpix/managers/preference_manager.dart';
 import 'package:kpix/tool_options/pencil_options.dart';
 import 'package:kpix/util/typedefs.dart';
-import 'package:kpix/widgets/main/layer_widget.dart';
 
 class PencilPainter extends IToolPainter
 {
