+++ lib/painting/shape_painter.dart
@@ -18,6 +18,8 @@ import 'dart:collection';
 import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
+import 'package:kpix/layer_states/drawing_layer_state.dart';
+import 'package:kpix/layer_states/layer_state.dart';
 import 'package:kpix/managers/hotkey_manager.dart';
 import 'package:kpix/preferences/behavior_preferences.dart';
 import 'package:kpix/util/helper.dart';
@@ -26,7 +28,6 @@ import 'package:kpix/painting/kpix_painter.dart';
 import 'package:kpix/managers/preference_manager.dart';
 import 'package:kpix/tool_options/shape_options.dart';
 import 'package:kpix/util/typedefs.dart';
-import 'package:kpix/widgets/main/layer_widget.dart';
 
 class ShapePainter extends IToolPainter
 {
