+++ lib/painting/stamp_painter.dart
@@ -17,6 +17,8 @@
 import 'dart:collection';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
+import 'package:kpix/layer_states/drawing_layer_state.dart';
+import 'package:kpix/layer_states/layer_state.dart';
 import 'package:kpix/util/helper.dart';
 import 'package:kpix/painting/itool_painter.dart';
 import 'package:kpix/painting/kpix_painter.dart';
@@ -24,7 +26,6 @@ import 'package:kpix/managers/preference_manager.dart';
 import 'package:kpix/managers/stamp_manager.dart';
 import 'package:kpix/tool_options/stamp_options.dart';
 import 'package:kpix/util/typedefs.dart';
-import 'package:kpix/widgets/main/layer_widget.dart';
 
 class StampPainter extends IToolPainter
 {
