+++ lib/util/typedefs.dart
@@ -17,11 +17,11 @@
 import 'dart:collection';
 
 
+import 'package:kpix/layer_states/layer_state.dart';
 import 'package:kpix/util/helper.dart';
 import 'package:kpix/widgets/file/import_widget.dart';
 import 'package:kpix/widgets/kpal/kpal_widget.dart';
 import 'package:kpix/widgets/file/export_widget.dart';
-import 'package:kpix/widgets/main/layer_widget.dart';
 
 typedef IdColorSelectedFn = void Function({required IdColor newColor});
 typedef ColorReferenceSelectedFn = void Function({required ColorReference? color});
