+++ lib/widgets/main/layer_widget.dart
@@ -14,580 +14,17 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import 'dart:async';
-import 'dart:collection';
-import 'dart:math';
-import 'dart:typed_data';
 import 'dart:ui' as ui;
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:get_it/get_it.dart';
+import 'package:kpix/layer_states/drawing_layer_state.dart';
+import 'package:kpix/layer_states/layer_state.dart';
+import 'package:kpix/layer_states/reference_layer_state.dart';
 import 'package:kpix/managers/hotkey_manager.dart';
-import 'package:kpix/managers/reference_image_manager.dart';
-import 'package:kpix/util/helper.dart';
-import 'package:kpix/widgets/kpal/kpal_widget.dart';
 import 'package:kpix/models/app_state.dart';
 import 'package:kpix/managers/preference_manager.dart';
-import 'package:kpix/util/typedefs.dart';
-import 'package:kpix/widgets/canvas/canvas_operations_widget.dart';
 import 'package:kpix/widgets/overlay_entries.dart';
-import 'package:kpix/widgets/tools/reference_layer_options_widget.dart';
-
-
-class LayerWidgetOptions
-{
-  final double outerPadding;
-  final double innerPadding;
-  final double borderRadius;
-  final double buttonSizeMin;
-  final double buttonSizeMax;
-  final double iconSize;
-  final double height;
-  final double dragOpacity;
-  final double borderWidth;
-  final double dragFeedbackSize;
-  final double dragTargetHeight;
-  final int dragTargetShowDuration;
-  final int dragDelay;
-  final int thumbUpdateTimerSec;
-  final int thumbUpdateTimerMsec;
-  final int addButtonSize;
-
-  LayerWidgetOptions({
-    required this.outerPadding,
-    required this.innerPadding,
-    required this.borderRadius,
-    required this.buttonSizeMin,
-    required this.buttonSizeMax,
-    required this.iconSize,
-    required this.height,
-    required this.dragOpacity,
-    required this.borderWidth,
-    required this.dragFeedbackSize,
-    required this.dragTargetHeight,
-    required this.dragTargetShowDuration,
-    required this.dragDelay,
-    required this.thumbUpdateTimerSec,
-    required this.thumbUpdateTimerMsec,
-    required this.addButtonSize});
-}
-
-enum LayerVisibilityState
-{
-  visible,
-  hidden
-}
-
-const Map<int, LayerVisibilityState> layerVisibilityStateValueMap =
-{
-  0: LayerVisibilityState.visible,
-  1: LayerVisibilityState.hidden,
-};
-
-enum LayerLockState
-{
-  unlocked,
-  transparency,
-  locked
-}
-
-const Map<int, LayerLockState> layerLockStateValueMap =
-{
-  0: LayerLockState.unlocked,
-  1: LayerLockState.transparency,
-  2: LayerLockState.locked
-};
-
-class ColorReference
-{
-  final KPalRampData ramp;
-  final int colorIndex;
-  ColorReference({required this.colorIndex, required this.ramp});
-  IdColor getIdColor()
-  {
-    return ramp.shiftedColors[colorIndex.clamp(0, ramp.shiftedColors.length - 1)].value;
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is ColorReference &&
-          runtimeType == other.runtimeType &&
-          ramp == other.ramp &&
-          colorIndex == other.colorIndex;
-
-  @override
-  int get hashCode => ramp.hashCode ^ colorIndex.hashCode;
-}
-
-abstract class LayerState
-{
-  final ValueNotifier<LayerVisibilityState> visibilityState = ValueNotifier(LayerVisibilityState.visible);
-  final ValueNotifier<bool> isSelected = ValueNotifier(false);
-  final ValueNotifier<ui.Image?> thumbnail = ValueNotifier(null);
-}
-
-class ReferenceLayerState extends LayerState
-{
-  final ReferenceLayerSettings _refSettings = GetIt.I.get<PreferenceManager>().referenceLayerSettings;
-  final ValueNotifier<int> opacityNotifier;
-  final ValueNotifier<double> aspectRatioNotifier;
-  final ValueNotifier<int> zoomNotifier;
-  final ValueNotifier<double> offsetXNotifier;
-  final ValueNotifier<double> offsetYNotifier;
-  final ValueNotifier<ReferenceImage?> imageNotifier;
-
-  ReferenceLayerState({required final int opacity, required final double aspectRatio, required final int zoom, required final ReferenceImage? image, required final double offsetX, required final double offsetY}) :
-    opacityNotifier = ValueNotifier(opacity),
-    aspectRatioNotifier = ValueNotifier(aspectRatio),
-    zoomNotifier = ValueNotifier(zoom),
-    offsetXNotifier = ValueNotifier(offsetX),
-    offsetYNotifier = ValueNotifier(offsetY),
-    imageNotifier = ValueNotifier(image)
-  {
-    if (image != null)
-    {
-      imageNotifier.value = image;
-      thumbnail.value = image.image;
-    }
-  }
-
-  factory ReferenceLayerState.from({required ReferenceLayerState other})
-  {
-    return ReferenceLayerState(aspectRatio: other.aspectRatioNotifier.value, opacity: other.opacityNotifier.value, zoom: other.zoomNotifier.value, image: other.imageNotifier.value, offsetX: other.offsetXNotifier.value, offsetY: other.offsetYNotifier.value);
-  }
-
-  void increaseZoom({final int step = 1})
-  {
-    final int newVal = zoomSliderValue + step;
-    setZoomSliderValue(newVal: newVal);
-  }
-
-  void decreaseZoom({final int step = 1})
-  {
-    final int newVal = zoomSliderValue - step;
-    setZoomSliderValue(newVal: newVal);
-  }
-
-  void setZoomSliderFromZoomFactor({required final double factor})
-  {
-    setZoomSliderValue(newVal: (pow(factor, 1.0 / _refSettings.zoomCurveExponent.toDouble()) * _refSettings.zoomDefault).round());
-  }
-
-  void setZoomSliderValue({required final int newVal})
-  {
-    if (newVal < _refSettings.zoomMin)
-    {
-      zoomNotifier.value = _refSettings.zoomMin;
-    }
-    else if (newVal > _refSettings.zoomMax)
-    {
-      zoomNotifier.value = _refSettings.zoomMax;
-    }
-    else
-    {
-      zoomNotifier.value = newVal;
-    }
-  }
-
-  int get opacity
-  {
-    return opacityNotifier.value;
-  }
-
-  double get aspectRatioFactorX
-  {
-    return (aspectRatioNotifier.value > 0) ? 1.0 + aspectRatioNotifier.value : 1.0;
-  }
-
-  double get aspectRatioFactorY
-  {
-    return (aspectRatioNotifier.value < 0) ? 1.0 - aspectRatioNotifier.value : 1.0;
-  }
-
-  int get zoomSliderValue
-  {
-    return zoomNotifier.value;
-  }
-
-  double get zoomFactor
-  {
-    return pow(zoomSliderValue.toDouble() / _refSettings.zoomDefault.toDouble(), _refSettings.zoomCurveExponent.toDouble()).toDouble();
-  }
-
-
-  double get offsetX
-  {
-    return offsetXNotifier.value;
-  }
-
-  double get offsetY
-  {
-    return offsetYNotifier.value;
-  }
-
-  ReferenceImage? get image
-  {
-    return imageNotifier.value;
-  }
-
-}
-
-class DrawingLayerState extends LayerState
-{
-  final ValueNotifier<LayerLockState> lockState = ValueNotifier(LayerLockState.unlocked);
-  final CoordinateSetI size;
-
-  final CoordinateColorMap _data;
-  ui.Image? raster;
-  bool isRasterizing = false;
-  bool doManualRaster = false;
-  final Map<CoordinateSetI, ColorReference?> rasterQueue = {};
-
-  DrawingLayerState._({required CoordinateColorMap data, required this.size, LayerLockState lState = LayerLockState.unlocked, LayerVisibilityState vState = LayerVisibilityState.visible}) : _data = data
-  {
-    isRasterizing = true;
-    _createRaster().then((final (ui.Image, ui.Image) images) => _rasterizingDone(image: images.$1, thb: images.$2, startedFromManual: false));
-    lockState.value = lState;
-    visibilityState.value = vState;
-    LayerWidgetOptions options = GetIt.I.get<PreferenceManager>().layerWidgetOptions;
-    Timer.periodic(Duration(seconds: options.thumbUpdateTimerSec, milliseconds: options.thumbUpdateTimerMsec), (final Timer t) {updateTimerCallback(timer: t);});
-
-  }
-
-  factory DrawingLayerState.from({required DrawingLayerState other})
-  {
-    final CoordinateColorMap data = HashMap();
-    for (final CoordinateColor ref in other._data.entries)
-    {
-      data[ref.key] = ref.value;
-    }
-    return DrawingLayerState._(size: other.size, data: data, lState: other.lockState.value, vState: other.visibilityState.value);
-  }
-
-  factory DrawingLayerState.deepClone({required DrawingLayerState other, required final KPalRampData originalRampData, required final KPalRampData rampData})
-  {
-    final CoordinateColorMap data = HashMap();
-    for (final CoordinateColor ref in other._data.entries)
-    {
-      ColorReference colRef = ref.value;
-
-      if (colRef.ramp == originalRampData)
-      {
-        colRef = rampData.references[colRef.colorIndex];
-      }
-      data[ref.key] = colRef;
-    }
-    return DrawingLayerState._(size: other.size, data: data, lState: other.lockState.value, vState: other.visibilityState.value);
-  }
-
-
-
-  factory DrawingLayerState({required CoordinateSetI size, final CoordinateColorMapNullable? content})
-  {
-    CoordinateColorMap data2 = HashMap();
-
-    if (content != null)
-    {
-      for (final CoordinateColorNullable entry in content.entries)
-      {
-        if (entry.key.x >= 0 && entry.key.y >= 0 && entry.key.x < size.x && entry.key.y < size.y && entry.value != null)
-        {
-          data2[entry.key] = entry.value!;
-        }
-      }
-    }
-    return DrawingLayerState._(data: data2, size: size);
-  }
-
-  void updateTimerCallback({required final Timer timer}) async
-  {
-    if ((rasterQueue.isNotEmpty || doManualRaster) && !isRasterizing)
-    {
-      isRasterizing = true;
-      _createRaster().then((final (ui.Image, ui.Image) images) => _rasterizingDone(image: images.$1, thb: images.$2, startedFromManual: doManualRaster));
-    }
-  }
-
-  void deleteRamp({required final KPalRampData ramp})
-  {
-    isRasterizing = true;
-    final Set<CoordinateSetI> deleteData = {};
-    for (final CoordinateColor entry in _data.entries)
-    {
-      if (entry.value.ramp == ramp)
-      {
-        deleteData.add(entry.key);
-      }
-    }
-
-    for (final CoordinateSetI coord in deleteData)
-    {
-      _data.remove(coord);
-    }
-
-    isRasterizing = false;
-  }
-
-  void remapAllColors({required final HashMap<ColorReference, ColorReference> rampMap})
-  {
-    isRasterizing = true;
-    for (final CoordinateColor entry in _data.entries)
-    {
-      _data[entry.key] = rampMap[entry.value]!;
-    }
-    isRasterizing = false;
-  }
-
-  void remapSingleRamp({required final KPalRampData newData, required final HashMap<int, int> map})
-  {
-    isRasterizing = true;
-    for (final CoordinateColor entry in _data.entries)
-    {
-      if (entry.value.ramp == newData)
-      {
-        _data[entry.key] = newData.references[map[entry.value.colorIndex]!];
-      }
-    }
-    isRasterizing = false;
-  }
-
-
-  void _rasterizingDone({required final ui.Image image, required final ui.Image thb, required final bool startedFromManual})
-  {
-    isRasterizing = false;
-    raster = image;
-    thumbnail.value = thb;
-    if (startedFromManual)
-    {
-      doManualRaster = false;
-    }
-    GetIt.I.get<AppState>().repaintNotifier.repaint();
-  }
-
-  Future<(ui.Image, ui.Image)> _createRaster() async
-  {
-    final AppState appState = GetIt.I.get<AppState>();
-    bool differentThumb = false;
-    if (appState.currentLayer == this && appState.selectionState.selection.hasValues())
-    {
-      differentThumb = true;
-    }
-
-    for (final CoordinateColorNullable entry in rasterQueue.entries)
-    {
-      if (entry.value == null)
-      {
-        _data.remove(entry.key);
-      }
-      else
-      {
-        _data[entry.key] = entry.value!;
-      }
-    }
-    rasterQueue.clear();
-
-    final ByteData byteDataImg = ByteData(size.x * size.y * 4);
-    ByteData? byteDataThumb;
-    if (differentThumb)
-    {
-      byteDataThumb = ByteData(size.x * size.y * 4);
-    }
-    for (final CoordinateColorNullable entry in _data.entries)
-    {
-      if (entry.value != null)
-      {
-        final Color dColor = entry.value!.getIdColor().color;
-        final int index = (entry.key.y * size.x + entry.key.x) * 4;
-        if (index < byteDataImg.lengthInBytes)
-        {
-          byteDataImg.setUint32(index, Helper.argbToRgba(argb: dColor.value));
-          if (byteDataThumb != null)
-          {
-            byteDataThumb.setUint32(index, Helper.argbToRgba(argb: dColor.value));
-          }
-        }
-      }
-    }
-    if (differentThumb)
-    {
-      final Iterable<CoordinateSetI> selectionCoords = appState.selectionState.selection.getCoordinates();
-      for (final CoordinateSetI coord in selectionCoords)
-      {
-        final ColorReference? colRef = appState.selectionState.selection.getColorReference(coord: coord);
-        if (colRef != null)
-        {
-          final Color dColor = colRef.getIdColor().color;
-          final int index = (coord.y * size.x + coord.x) * 4;
-          if (index > 0 && index < byteDataThumb!.lengthInBytes)
-          {
-            byteDataThumb.setUint32(index, Helper.argbToRgba(argb: dColor.value));
-          }
-        }
-      }
-    }
-
-
-    final Completer<ui.Image> completerImg = Completer<ui.Image>();
-    ui.decodeImageFromPixels(
-      byteDataImg.buffer.asUint8List(),
-      size.x,
-      size.y,
-      ui.PixelFormat.rgba8888, (ui.Image convertedImage)
-      {
-        completerImg.complete(convertedImage);
-      }
-    );
-
-    final ui.Image img = await completerImg.future;
-
-    if (differentThumb)
-    {
-      final Completer<ui.Image> completerThumb = Completer<ui.Image>();
-      ui.decodeImageFromPixels(
-          byteDataThumb!.buffer.asUint8List(),
-          size.x,
-          size.y,
-          ui.PixelFormat.rgba8888, (ui.Image convertedImage)
-      {
-        completerThumb.complete(convertedImage);
-      }
-      );
-      final ui.Image thumb = await completerThumb.future;
-      return (img, thumb);
-    }
-    else
-    {
-      return (img, img);
-    }
-  }
-
-
-  ColorReference? getDataEntry({required final CoordinateSetI coord})
-  {
-    if (_data.containsKey(coord))
-    {
-      return _data[coord];
-    }
-    else if (rasterQueue.isNotEmpty)
-    {
-      return rasterQueue[coord];
-    }
-    return null;
-  }
-
-  CoordinateColorMap getData()
-  {
-    return _data;
-  }
-
-  void setDataAll({required final CoordinateColorMapNullable list})
-  {
-    rasterQueue.addAll(list);
-  }
-
-  Future <void> removeDataAll({required final Set<CoordinateSetI> removeCoordList}) async
-  {
-    for (final CoordinateSetI coord in removeCoordList)
-    {
-     rasterQueue[coord] = null;
-    }
-  }
-
-
-  DrawingLayerState getTransformedLayer({required final CanvasTransformation transformation})
-  {
-    final CoordinateColorMap rotatedContent = HashMap();
-    final CoordinateSetI newSize = CoordinateSetI.from(other: size);
-    if (transformation == CanvasTransformation.rotate)
-    {
-      newSize.x = size.y;
-      newSize.y = size.x;
-    }
-    for (final CoordinateColor entry in _data.entries)
-    {
-      final CoordinateSetI rotCoord = CoordinateSetI.from(other: entry.key);
-      if (transformation == CanvasTransformation.rotate)
-      {
-        rotCoord.x = ((size.y - 1) - entry.key.y).toInt();
-        rotCoord.y = entry.key.x;
-      }
-      else if (transformation == CanvasTransformation.flipH)
-      {
-        rotCoord.x = ((size.x - 1) - entry.key.x).toInt();
-      }
-      else if (transformation == CanvasTransformation.flipV)
-      {
-        rotCoord.y = ((size.y - 1) - entry.key.y).toInt();
-      }
-
-      rotatedContent[rotCoord] = entry.value;
-    }
-    if (rasterQueue.isNotEmpty)
-    {
-      for (final CoordinateColorNullable entry in rasterQueue.entries)
-      {
-        final CoordinateSetI rotCoord = CoordinateSetI.from(other: entry.key);
-        if (transformation == CanvasTransformation.rotate)
-        {
-          rotCoord.x = ((size.y - 1) - entry.key.y).toInt();
-          rotCoord.y = entry.key.x;
-        }
-        else if (transformation == CanvasTransformation.flipH)
-        {
-          rotCoord.x = ((size.x - 1) - entry.key.x).toInt();
-        }
-        else if (transformation == CanvasTransformation.flipV)
-        {
-          rotCoord.y = ((size.y - 1) - entry.key.y).toInt();
-        }
-        if (entry.value != null)
-        {
-          rotatedContent[rotCoord] = entry.value!;
-        }
-        else if (rotatedContent.containsKey(rotCoord))
-        {
-          rotatedContent.remove(rotCoord);
-        }
-      }
-    }
-    return DrawingLayerState(size: newSize, content: rotatedContent);
-  }
-
-  DrawingLayerState getResizedLayer({required final CoordinateSetI newSize, required final CoordinateSetI offset})
-  {
-    final CoordinateColorMap croppedContent = HashMap();
-    for (final CoordinateColor entry in _data.entries)
-    {
-      final CoordinateSetI newCoord = CoordinateSetI(x: entry.key.x + offset.x, y: entry.key.y + offset.y);
-      if (newCoord.x >= 0 && newCoord.x < newSize.x && newCoord.y >= 0 && newCoord.y < newSize.y)
-      {
-        croppedContent[newCoord] = entry.value;
-      }
-    }
-
-    if (rasterQueue.isNotEmpty)
-    {
-      for (final CoordinateColorNullable entry in rasterQueue.entries)
-      {
-        final CoordinateSetI newCoord = CoordinateSetI(x: entry.key.x + offset.x, y: entry.key.y + offset.y);
-        if (newCoord.x >= 0 && newCoord.x < newSize.x && newCoord.y >= 0 && newCoord.y < newSize.y)
-        {
-          if (entry.value != null)
-          {
-            croppedContent[newCoord] = entry.value!;
-          }
-          else if (croppedContent.containsKey(newCoord))
-          {
-            croppedContent.remove(newCoord);
-          }
-        }
-      }
-    }
-    return DrawingLayerState(size: newSize, content: croppedContent);
-  }
-}
-
 
 class LayerWidget extends StatefulWidget
 {
@@ -598,8 +35,7 @@ class LayerWidget extends StatefulWidget
   });
 
   @override
-  State<LayerWidget> createState() => _LayerWidgetState();  
-  
+  State<LayerWidget> createState() => _LayerWidgetState();
 }
 
 class _LayerWidgetState extends State<LayerWidget>
